Share the authz environment through a package conftest
test_generic_field_reader_allowlist.py borrowed the #125 suite's fixtures by
module-attribute copy, so `env` growing a `_module_env` dependency broke every
test in it at setup. Both suites now live in tests/multi_project_authz/ and take
`env` from the package conftest, where dependencies resolve themselves and
autouse stays scoped to exactly those two files.

Test discovery in the CI classification guardrail recurses, or a sub-package
could drop out of the gate without going red.
