+++ .github/workflows/ci.yml
@@ -305,6 +305,8 @@ jobs:
           CI_SHARD: ${{ matrix.shard }}
         run: >-
           python -m pytest $PYTEST_FLAGS --ci-shard "$CI_SHARD/8"
+          tests/multi_project_authz/test_generic_field_reader_allowlist.py
+          tests/multi_project_authz/test_multi_project_membership_authz.py
           tests/test_adapter_attachment.py
           tests/test_adapter_header.py
           tests/test_aitoolkit_run.py
@@ -345,7 +347,6 @@ jobs:
           tests/test_finder_query_cost.py
           tests/test_florence.py
           tests/test_florence_model_variant.py
-          tests/test_generic_field_reader_allowlist.py
           tests/test_hub_bootstrap.py
           tests/test_hub_engine.py
           tests/test_hub_model_shelf_schema.py
@@ -375,7 +376,6 @@ jobs:
           tests/test_model_shelf_library_fixture.py
           tests/test_model_shelf_scan_invariants.py
           tests/test_model_unload_race.py
-          tests/test_multi_project_membership_authz.py
           tests/test_openapi_response_schemas.py
           tests/test_operation_log.py
           tests/test_path_containment.py
@@ -706,7 +706,7 @@ jobs:
           tests/test_snapshots_auth.py
           tests/test_metadata_hash.py
           tests/test_architecture_guardrails.py
-          tests/test_multi_project_membership_authz.py
+          tests/multi_project_authz/test_multi_project_membership_authz.py
           tests/test_openapi_response_schemas.py
           tests/test_find_orphaned_files.py
           tests/test_server.py
