+++ tests/ci_test_durations.json
@@ -1,5 +1,29 @@
 {
   "durations": {
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_add_member_to_shared_set_joins_every_project": 10.01,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_bulk_add_to_shared_set_joins_every_project": 11.18,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_bulk_replace_members_joins_every_project": 10.45,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_character_by_id_secondary_project_token_reaches_it": 11.96,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_character_by_project_and_name_both_directions": 11.86,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_character_list_is_narrowed_to_the_tokens_project": 10.69,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_deleting_a_project_leaves_the_other_membership_intact": 10.39,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_face_assignment_to_shared_character_joins_every_project": 10.95,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_import_into_shared_character_joins_every_project": 12.28,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_import_into_shared_set_joins_every_project": 7.97,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_leaving_one_project_keeps_the_other": 10.08,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_locked_members_listing_both_directions": 9.71,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_membership_is_written_to_both_representations": 6.59,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_owner_project_filters_read_the_join": 7.45,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_picture_scope_follows_the_shared_set": 9.38,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_picture_set_by_id_secondary_project_token_reaches_it": 10.37,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_picture_set_by_project_and_name_both_directions": 10.63,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_picture_set_list_is_narrowed_to_the_tokens_project": 5.99,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_project_ids_is_empty_for_entity_scoped_tokens": 10.55,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_project_ids_narrowed_to_the_tokens_own_project": 9.75,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_project_picture_sets_listing_both_directions": 10.25,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_scalar_project_id_is_derived_from_the_narrowed_list": 9.32,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_scoped_list_pictures_not_over_blocked": 10.89,
+    "tests/multi_project_authz/test_multi_project_membership_authz.py::test_unknown_project_id_is_404_not_a_silent_partial_write": 12.29,
     "tests/test_anomaly_region_scope.py::test_anomaly_region_cache_hit_is_scope_gated": 12.83,
     "tests/test_anomaly_region_scope.py::test_anomaly_region_character_scope_both_directions": 12.93,
     "tests/test_anomaly_region_scope.py::test_anomaly_region_missing_picture_is_404": 12.67,
@@ -935,30 +959,6 @@
     "tests/test_migrations.py::test_the_migration_chain_has_exactly_one_head": 0.56,
     "tests/test_migrations.py::test_token_reset_clears_tokens_guest_rows_and_stored_addresses": 2.03,
     "tests/test_migrations.py::test_token_reset_runs_on_already_empty_tables": 3.43,
-    "tests/test_multi_project_membership_authz.py::test_add_member_to_shared_set_joins_every_project": 10.01,
-    "tests/test_multi_project_membership_authz.py::test_bulk_add_to_shared_set_joins_every_project": 11.18,
-    "tests/test_multi_project_membership_authz.py::test_bulk_replace_members_joins_every_project": 10.45,
-    "tests/test_multi_project_membership_authz.py::test_character_by_id_secondary_project_token_reaches_it": 11.96,
-    "tests/test_multi_project_membership_authz.py::test_character_by_project_and_name_both_directions": 11.86,
-    "tests/test_multi_project_membership_authz.py::test_character_list_is_narrowed_to_the_tokens_project": 10.69,
-    "tests/test_multi_project_membership_authz.py::test_deleting_a_project_leaves_the_other_membership_intact": 10.39,
-    "tests/test_multi_project_membership_authz.py::test_face_assignment_to_shared_character_joins_every_project": 10.95,
-    "tests/test_multi_project_membership_authz.py::test_import_into_shared_character_joins_every_project": 12.28,
-    "tests/test_multi_project_membership_authz.py::test_import_into_shared_set_joins_every_project": 7.97,
-    "tests/test_multi_project_membership_authz.py::test_leaving_one_project_keeps_the_other": 10.08,
-    "tests/test_multi_project_membership_authz.py::test_locked_members_listing_both_directions": 9.71,
-    "tests/test_multi_project_membership_authz.py::test_membership_is_written_to_both_representations": 6.59,
-    "tests/test_multi_project_membership_authz.py::test_owner_project_filters_read_the_join": 7.45,
-    "tests/test_multi_project_membership_authz.py::test_picture_scope_follows_the_shared_set": 9.38,
-    "tests/test_multi_project_membership_authz.py::test_picture_set_by_id_secondary_project_token_reaches_it": 10.37,
-    "tests/test_multi_project_membership_authz.py::test_picture_set_by_project_and_name_both_directions": 10.63,
-    "tests/test_multi_project_membership_authz.py::test_picture_set_list_is_narrowed_to_the_tokens_project": 5.99,
-    "tests/test_multi_project_membership_authz.py::test_project_ids_is_empty_for_entity_scoped_tokens": 10.55,
-    "tests/test_multi_project_membership_authz.py::test_project_ids_narrowed_to_the_tokens_own_project": 9.75,
-    "tests/test_multi_project_membership_authz.py::test_project_picture_sets_listing_both_directions": 10.25,
-    "tests/test_multi_project_membership_authz.py::test_scalar_project_id_is_derived_from_the_narrowed_list": 9.32,
-    "tests/test_multi_project_membership_authz.py::test_scoped_list_pictures_not_over_blocked": 10.89,
-    "tests/test_multi_project_membership_authz.py::test_unknown_project_id_is_404_not_a_silent_partial_write": 12.29,
     "tests/test_openapi_response_schemas.py::test_no_endpoint_has_null_success_response": 3.09,
     "tests/test_openapi_response_schemas.py::test_no_operation_has_duplicate_tags": 3.65,
     "tests/test_openapi_response_schemas.py::test_no_query_param_declares_a_default": 2.82,
