+++ tests/multi_project_authz/__init__.py
@@ -1,0 +1,1 @@
+"""Shared environment for the multi-project membership authz suites."""
